scripts fix

+++ src/migrations/2015_06_09_033732_create_blusers_table.php
@@ -13,7 +13,10 @@ class CreateBLUsersTable extends Migration {
 	public function up()
 	{
 		//first we need to remove the standard laravel one:
-		Schema::drop('users');
+		if (Schema::hasTable('users'))
+		{
+			Schema::drop('users');
+		}
 		
 		Schema::create('users', function(Blueprint $table)
 		{
